refactor: extract resolver ownership
